refactor: ultimate refactoring of recordScreen
