+++ windrecorder/config.py
@@ -39,7 +39,7 @@ class Config:
         video_compress_rate,
         oneday_timeline_pic_num,
         enable_ocr_chineseocr_lite_onnx,
-        maintain_lock_path="cache\\LOCK_MAINTAIN.MD",
+        maintain_lock_path="cache\\LOCK_MAINTAIN",
         record_lock_path="cache\\LOCK_FILE_RECORD.MD",
         last_idle_maintain_file_path="cache\\LAST_IDLE_MAINTAIN.MD",
         **other_field,
